Move client's options to separate files
This should help in discovering the available options for a given object.
