+++ spec_opts.go
@@ -1,0 +1,14 @@
+package containerd
+
+import specs "github.com/opencontainers/runtime-spec/specs-go"
+
+// SpecOpts sets spec specific information to a newly generated OCI spec
+type SpecOpts func(s *specs.Spec) error
+
+// WithProcessArgs replaces the args on the generated spec
+func WithProcessArgs(args ...string) SpecOpts {
+	return func(s *specs.Spec) error {
+		s.Process.Args = args
+		return nil
+	}
+}
