+++ spec_windows.go
@@ -1,16 +1,6 @@
 package containerd
 
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-
-	"github.com/containerd/containerd/containers"
-	"github.com/containerd/containerd/images"
-	"github.com/containerd/containerd/typeurl"
-	"github.com/opencontainers/image-spec/specs-go/v1"
-	specs "github.com/opencontainers/runtime-spec/specs-go"
-)
+import specs "github.com/opencontainers/runtime-spec/specs-go"
 
 func createDefaultSpec() (*specs.Spec, error) {
 	return &specs.Spec{
@@ -30,69 +20,4 @@ func createDefaultSpec() (*specs.Spec, error) {
 			},
 		},
 	}, nil
-}
-
-func WithImageConfig(ctx context.Context, i Image) SpecOpts {
-	return func(s *specs.Spec) error {
-		var (
-			image = i.(*image)
-			store = image.client.ContentStore()
-		)
-		ic, err := image.i.Config(ctx, store)
-		if err != nil {
-			return err
-		}
-		var (
-			ociimage v1.Image
-			config   v1.ImageConfig
-		)
-		switch ic.MediaType {
-		case v1.MediaTypeImageConfig, images.MediaTypeDockerSchema2Config:
-			r, err := store.Reader(ctx, ic.Digest)
-			if err != nil {
-				return err
-			}
-			if err := json.NewDecoder(r).Decode(&ociimage); err != nil {
-				r.Close()
-				return err
-			}
-			r.Close()
-			config = ociimage.Config
-		default:
-			return fmt.Errorf("unknown image config media type %s", ic.MediaType)
-		}
-		s.Process.Env = config.Env
-		s.Process.Args = append(config.Entrypoint, config.Cmd...)
-		s.Process.User = specs.User{
-			Username: config.User,
-		}
-		return nil
-	}
-}
-
-func WithTTY(width, height int) SpecOpts {
-	return func(s *specs.Spec) error {
-		s.Process.Terminal = true
-		s.Process.ConsoleSize.Width = uint(width)
-		s.Process.ConsoleSize.Height = uint(height)
-		return nil
-	}
-}
-
-func WithSpec(spec *specs.Spec) NewContainerOpts {
-	return func(ctx context.Context, client *Client, c *containers.Container) error {
-		any, err := typeurl.MarshalAny(spec)
-		if err != nil {
-			return err
-		}
-		c.Spec = any
-		return nil
-	}
-}
-
-func WithResources(resources *specs.WindowsResources) UpdateTaskOpts {
-	return func(ctx context.Context, client *Client, r *UpdateTaskInfo) error {
-		r.Resources = resources
-		return nil
-	}
 }
